+++ Dyna/Cards/BunkerCard.cs
@@ -19,40 +19,17 @@ internal sealed class BunkerCard : Card, IRegisterable
 				upgradesTo = [Upgrade.A, Upgrade.B]
 			},
 			Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/Bunker.png")).Sprite,
-			Name = ModEntry.Instance.AnyLocalizations.Bind(["card", "Bunker", "name"]).Localize
+			Name = ModEntry.Instance.AnyLocalizations.Bind(["card", "Bunker", "name"]).Localize,
 		});
 	}
 
 	public override CardData GetData(State state)
-		=> new()
-		{
-			cost = upgrade == Upgrade.B ? 0 : 2
-		};
+		=> new() { cost = upgrade == Upgrade.B ? 0 : 2 };
 
 	public override List<CardAction> GetActions(State s, Combat c)
 		=> [
-			new AStatus
-			{
-				targetPlayer = true,
-				status = Status.shield,
-				statusAmount = upgrade switch
-				{
-					Upgrade.A => 3,
-					Upgrade.B => 1,
-					_ => 2
-				}
-			},
-			new AStatus
-			{
-				targetPlayer = true,
-				status = Status.tempShield,
-				statusAmount = 1
-			},
-			new AStatus
-			{
-				targetPlayer = true,
-				status = upgrade == Upgrade.B ? Status.energyLessNextTurn : Status.energyNextTurn,
-				statusAmount = 1
-			}
+			new AStatus { targetPlayer = true, status = Status.shield, statusAmount = upgrade == Upgrade.A ? 3 : 2 },
+			new AStatus { targetPlayer = true, status = Status.tempShield, statusAmount = 1 },
+			new AStatus { targetPlayer = true, status = upgrade == Upgrade.B ? Status.energyLessNextTurn : Status.energyNextTurn, statusAmount = 1 },
 		];
 }